requirements.txt
# Core Dependencies
openai>=1.50.0
google-generativeai>=0.8.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
backoff>=2.2.1

# Web Search and Content Processing
brave-search>=0.1.8
beautifulsoup4>=4.12.0
requests>=2.31.0
httpx>=0.25.0

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
nltk>=3.8.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0

# Caching and Storage
cachetools>=5.3.0
joblib>=1.3.0

# Logging and Monitoring
structlog>=23.1.0

# Testing (Optional - for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality (Optional - for development)
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Web Framework (Optional - for future API)
fastapi>=0.100.0
uvicorn>=0.23.0

# Knowledge Graph (Optional)
rdflib>=7.0.0
sparqlwrapper>=2.0.0 